refactor(docs): Welcome のリンクマニフェストを非 stories モジュールへ切り出し (#399)
* refactor(docs): Welcome のリンクマニフェストを非 stories モジュールへ切り出し

`WELCOME_DEEP_LINKS` / `WELCOME_COMPONENT_SLUGS` は `Welcome.stories.tsx` の
named export だったため、Storybook の CSF スキャナがストーリーと誤認し
`WELCOME DEEP LINKS` / `WELCOME COMPONENT SLUGS` という空のサイドバー項目を
2つ生成していた。

定数と `WelcomeDeepLink` 型をプレーンな `Welcome.links.ts` へ移動し、
stories と drift テストの双方がそこから import するように変更。CSF は
`*.stories.tsx` しか走査しないため、サイドバー漏れが構造的に解消する
(`excludeStories` のような後付け抑制が不要)。

- index.json の Welcome 配下が welcome--overview の1件のみに (以前3件)
- Welcome ページの deep-link カード27枚は href 込みで従来どおり描画
- Welcome.drift.test.ts 56 passed / typecheck / lint いずれもクリーン

Storybook 内部のみの変更で公開 surface (.st-* / props / CSS変数) に影響しない
ため changeset は不要 (story IDs are internal)。


* chore(changeset): add patch changeset for Welcome links extraction

Storybook 内部のみの変更だが、本リポジトリの changeset gate は
`changeset status --since=origin/<base>` で空変更を許さないため、
docs-only PR の慣例 (#377 等) に倣い patch + 公開 API 不変の明記とする。


---------
